requirements.txt
asgiref==3.8.1
Django==4.2.17
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.2
typing-extensions==4.12.2
